requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.3
pandas==1.1.4
pefile==2019.4.18
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
PyJWT==2.1.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.4
pywin32-ctypes==0.2.0
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.11
virtualenv==20.2.1
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
